requirements.txt
Django>=4.2,<5
django-environ>=0.12
djangorestframework>=3.15
djangorestframework-simplejwt>=5.5
channels>=4.0